Add cursor_id to Cursor
